+++ DLVN-SDK/src/main/java/com/edoctor/dlvn_sdk/api/ApiService.kt
@@ -34,5 +34,11 @@ interface ApiService {
     fun expireEClinicRinging(@Header("Authorization") token: String, @Body params: JsonObject): Call<Any>
 
     @POST("graphql")
+    fun cancelAppointmentSchedule(@Header("Authorization") token: String, @Body params: JsonObject): Call<Any>
+
+    @POST("graphql")
+    fun confirmAppointmentSchedule(@Header("Authorization") token: String, @Body params: JsonObject): Call<Any>
+
+    @POST("graphql")
     fun getAppointmentDetail(@Body params: JsonObject): Call<AppointmentDetailResponse>
 }