+++ DLVN-SDK/src/main/java/com/edoctor/dlvn_sdk/sendbirdCall/WebViewCallActivity.kt
@@ -2,6 +2,7 @@ package com.edoctor.dlvn_sdk.sendbirdCall
 
 import android.annotation.SuppressLint
 import android.os.Bundle
+import android.util.Log
 import android.view.View
 import android.view.WindowManager
 import android.widget.ImageButton
@@ -62,6 +63,7 @@ class WebViewCallActivity: AppCompatActivity() {
         btnToggleMic = findViewById(R.id.btn_toggle_mic_pip)
         btnToggleCam = findViewById(R.id.btn_toggle_cam_pip)
         btnEndCall!!.setOnClickListener {
+            callManager?.endEclinicCall()
             directCall?.end()
         }
         btnToggleCam!!.setOnClickListener {