+++ DLVN-SDK/src/main/java/com/edoctor/dlvn_sdk/store/AppStore.kt
@@ -3,6 +3,7 @@ package com.edoctor.dlvn_sdk.store
 import android.annotation.SuppressLint
 import com.edoctor.dlvn_sdk.EdoctorDlvnSdk
 import com.edoctor.dlvn_sdk.webview.SdkWebView
+import com.edoctor.dlvn_sdk.widget.AppointmentListAdapter
 
 object AppStore {
     @SuppressLint("StaticFieldLeak")
@@ -11,4 +12,6 @@ object AppStore {
     var sdkInstance: EdoctorDlvnSdk? = null
     var activeChannelUrl: String? = null
     var isAppInForeground: Boolean = false
+    var widgetList: AppointmentListAdapter? = null
+    var updateWidgetListDisplay: ((state: String) -> Unit)? = {}
 }