+++ app/src/main/java/com/edoctor/application/MainActivity.kt
@@ -1,12 +1,11 @@
 package com.edoctor.application
 
+import android.content.res.Resources
 import android.os.Bundle
 import android.util.Log
 import android.view.Gravity
 import android.view.View
-import android.view.ViewGroup
 import android.widget.Button
-import android.widget.ImageView
 import android.widget.TextView
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity